Added new activity for text edition Added text edition to NoteViewerActivity Added new query to DataDAO Changes in layouts
