+++ app/src/main/AndroidManifest.xml
@@ -14,6 +14,8 @@
 
         <activity android:name=".activity.TemplateEmptyActivity"
             />
+        <activity android:name=".activity.TextEditorActivityOld"
+            />
         <activity android:name=".activity.TextEditorActivity"
             />
         <activity android:name=".activity.ListActivity"
