+++ app/src/main/java/com/thesis/note/database/dao/DataDAO.kt
@@ -2,6 +2,7 @@ package com.thesis.note.database.dao
 
 import androidx.room.*
 import com.thesis.note.database.entity.Data
+import com.thesis.note.database.entity.Note
 
 @Dao
 interface DataDAO {
@@ -11,7 +12,8 @@ interface DataDAO {
     @Query("SELECT * FROM data WHERE NoteId = (:noteID)")
     fun getDataFromNote(noteID:Int):List<Data>
 
-
+    @Query("SELECT * FROM data WHERE IdData = (:noteID)")
+    fun getDataById(noteID:Int): Data
 
     @Insert
     fun insertAll(vararg insert: Data): List<Long>
